refactor(MovieDetail): 별점 mouseover 이벤트 내부 로직으로 수정
